Finish porting all settings. Make the sync manager use the new setup
