+++ src/main/java/de/tr7zw/firstperson/features/layers/HeadLayerFeatureRenderer.java
@@ -81,15 +81,15 @@ public class HeadLayerFeatureRenderer
 	private static MinecraftClient client = MinecraftClient.getInstance();
 	
 	public static boolean isEnabled(AbstractClientPlayerEntity abstractClientPlayerEntity) {
-		LayerMode mode = FirstPersonModelMod.config.layerMode;
+		LayerMode mode = FirstPersonModelMod.config.skinLayer.headLayerMode;
 		if (mode == LayerMode.VANILLA2D || !abstractClientPlayerEntity.isPartVisible(PlayerModelPart.HAT))
 			return false;
 		if (client.player == abstractClientPlayerEntity) {
 			return true;
 		}
 		if (mode != LayerMode.ONLYSELF) {
-			int distance = FirstPersonModelMod.config.optimizedLayerDistance
-					* FirstPersonModelMod.config.optimizedLayerDistance;
+			int distance = FirstPersonModelMod.config.skinLayer.optimizedLayerDistance
+					* FirstPersonModelMod.config.skinLayer.optimizedLayerDistance;
 			boolean ret = client.player.getPos().squaredDistanceTo(abstractClientPlayerEntity.getPos()) < distance;
 			return ret;
 		}
